src/api_ia/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyodbc==4.0.39
pandas==2.1.3
numpy==1.26.2
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
scikit-learn==1.3.2 